requirements-test.txt
pytest
pytest-playwright
pyyaml
